Rename config.Configuration to config.Config
Now that we got rid of the complicated viper wrapper, we
can name this Config, which is a bit more idiomatic.
